Lots of fixes
W/o a control flow graph we really hand a ton of foot guns to the programmer. For instance, a static data member that references an object instance becomes live because we default initialize it.

Returns are optional. But if they are present we have to ensure that their type is correct. Consider the following:

class A {
  public int foo() {
    if (false) {
      return 4;
    }
   //No way to tell that all paths return w/o control flow.
  }
}